perf(hpo): persist NF4 base model across trials + HF offline mode
Optuna's study.optimize runs trials in-process, but each trial's
train_and_register and _evaluate_adapter_on_heldout each called
AutoModelForCausalLM.from_pretrained from scratch. On a 30-trial study
that's 60 fresh NF4 loads of Qwen3.5-9B (~3 min each on cold cache,
~30-60 s on warm OS page cache). Empirical 2-trial smoke confirms full
module-skip wins ~2.5 min per cached trial — extrapolates to ~72 min
saved on a 30-trial study.

Architecture:
- New module-level _BASE_MODEL_CACHE keyed by model_id, gated on
  RUNE_PERSIST_BASE_MODEL=1 (set by run_hpo.sh).
- _get_or_load_base() returns cached (model, tokenizer) or loads fresh.
  Used by trainer.train_qlora and run_training_hpo._evaluate_adapter_on_heldout.
- _release_trial_state() at the end of train_qlora calls
  trainer.model.unload() (PeftModel) to remove LoRA layers in-place,
  restoring the cached base for the next trial. Falls back to plain del
  + gc + torch.cuda.empty_cache when persistence is disabled.
- Eval path mirrors: PeftModel.from_pretrained(cached_base, adapter_path)
  then .unload() after the forward passes so the next trial's base is
  clean.
- Skeptic-approved freebie: del trainer + gc + torch.cuda.empty_cache
  always runs at end of train_qlora, breaks the train+eval co-residency
  that pushed the L4 close to 22GB peak.

HF offline:
- run_hpo.sh now exports HF_HUB_OFFLINE=1 and TRANSFORMERS_OFFLINE=1.
  Skips ~6-8 cache-validation HEAD checks per trial. Marginal time,
  but removes flaky-network failure modes mid-study.

Drive-by:
- run_training_hpo.py:_evaluate_adapter_on_heldout: dropped torch_dtype=
  kwarg (deprecated for dtype=) and the explicit param entirely since
  bf16 compute dtype is already in BitsAndBytesConfig.
- experiment_harness.py: torch_dtype= → dtype= for the same deprecation.

Empirical from 2-trial smoke (data/pairs/fastapi_fastapi.jsonl):
- Trial 0: ~3 min (cold load + train + eval)
- Trial 1: 22 seconds (cache hit + train 11s + eval 11s)
- Cache reuse logged 3 times, both trials completed, fitness=0.428

+++ scripts/experiment_harness.py
@@ -74,7 +74,7 @@ def setup() -> None:
     dtype = resolve_model_dtype(param_count=param_count, device=DEVICE)
     print(f"Inference dtype: {dtype}")
 
-    _model = AutoModelForCausalLM.from_pretrained(MODEL_NAME, torch_dtype=dtype)
+    _model = AutoModelForCausalLM.from_pretrained(MODEL_NAME, dtype=dtype)
     _model.to(DEVICE)  # type: ignore[arg-type]
     _model.eval()
     print("Model loaded.\n")
